requirements.txt
lightning
matplotlib
addict
fire
pyyaml
toml
tensorboardX